Reorganize contact and distance map functionality
