+++ js/ensembleManager.js
@@ -1,13 +1,10 @@
 import * as THREE from "../node_modules/three/build/three.module.js";
 import Globals from './globals.js';
 import igv from '../vendor/igv.esm.js'
-import KDBush from '../node_modules/kd3d/js/index.js'
-import { rgb255String, appleCrayonColorRGB255 } from './color.js';
 import { distanceMapPanel, contactFrequencyMapPanel } from './gui.js';
+import { getEnsembleAverageDistanceMapCanvas } from './distanceMapPanel.js';
+import { getEnsembleContactFrequencyCanvas } from './contactFrequencyMapPanel.js';
 import { numberFormatter, readFileAsText } from "./utils.js";
-
-export let contactFrequencyDistanceThreshold = 256;
-// export let contactFrequencyDistanceThreshold = 3 * 1024;
 
 class EnsembleManager {
 
@@ -109,7 +106,7 @@ class EnsembleManager {
         // segmentIDSanityCheck(this.ensemble);
 
         // update ensemble level distance map
-        // distanceMapPanel.drawEnsembleDistanceCanvas(getEnsembleDistanceMapCanvas(this.ensemble));
+        distanceMapPanel.drawEnsembleDistanceCanvas(getEnsembleAverageDistanceMapCanvas(this.ensemble));
 
         const { chr, genomicStart, genomicEnd } = this.locus;
 
@@ -225,300 +222,6 @@ const segmentIDSanityCheck = ensemble => {
     }
 
     console.timeEnd(`segmentIDSanityCheck. ${ ensembleList.length } traces.`);
-
-};
-
-export const getEnsembleContactFrequencyCanvas = (ensemble, distanceThreshold) => {
-
-    let mapSize = Globals.ensembleManager.maximumSegmentID;
-
-    let frequencies = new Array(mapSize * mapSize);
-    for (let f = 0; f < frequencies.length; f++) frequencies[ f ] = 0;
-
-    const ensembleList = Object.values(ensemble);
-    console.time(`getEnsembleContactFrequencyCanvas. ${ ensembleList.length } traces.`);
-
-    for (let trace of ensembleList) {
-        updateContactFrequencyArray(trace, frequencies, distanceThreshold);
-    }
-
-    console.timeEnd(`getEnsembleContactFrequencyCanvas. ${ ensembleList.length } traces.`);
-
-    return createContactFrequencyMap(frequencies);
-
-};
-
-export const getTraceContactFrequencyCanvas = (trace, distanceThreshold) => {
-
-    let mapSize = Globals.ensembleManager.maximumSegmentID;
-
-    let frequencies = new Array(mapSize * mapSize);
-    for (let f = 0; f < frequencies.length; f++) frequencies[ f ] = 0;
-
-    updateContactFrequencyArray(trace, frequencies, distanceThreshold);
-
-    return createContactFrequencyMap(frequencies);
-
-};
-
-const updateContactFrequencyArray = (trace, frequencies, distanceThreshold) => {
-
-    let { vertices } = trace.geometry;
-    let { segmentList } = trace;
-    const exclusionSet = new Set();
-
-    const spatialIndex = new KDBush(kdBushConfiguratorWithTrace(trace));
-
-    const mapSize = Globals.ensembleManager.maximumSegmentID;
-
-    for (let i = 0; i < vertices.length; i++) {
-
-        const { x, y, z } = vertices[ i ];
-
-        exclusionSet.add(i);
-
-        const xy_diagonal = (segmentList[ i ].segmentID - 1) * mapSize + (segmentList[ i ].segmentID - 1);
-        frequencies[ xy_diagonal ]++;
-
-        const contact_indices = spatialIndex.within(x, y, z, distanceThreshold).filter(index => !exclusionSet.has(index));
-
-        if (contact_indices.length > 0) {
-            for (let contact_i of contact_indices) {
-
-                const         i_frequency = segmentList[         i ].segmentID - 1;
-                const contact_i_frequency = segmentList[ contact_i ].segmentID - 1;
-
-                const xy =         i_frequency * mapSize + contact_i_frequency;
-                const yx = contact_i_frequency * mapSize +         i_frequency;
-
-                if (xy > frequencies.length) {
-                    console.log('xy is bogus index ' + xy);
-                }
-
-                if (yx > frequencies.length) {
-                    console.log('yx is bogus index ' + yx);
-                }
-
-                ++frequencies[ xy ];
-
-                frequencies[ yx ] = frequencies[ xy ];
-
-            }
-        }
-
-    }
-
-};
-
-const createContactFrequencyMap = frequencies => {
-
-    let mapSize = Globals.ensembleManager.maximumSegmentID;
-
-    console.time(`Create ${ mapSize } x ${ mapSize } contact map`);
-
-    let maxFrequency = Number.NEGATIVE_INFINITY;
-
-    // Calculate max
-    for (let m = 0; m < mapSize; m++) {
-        const xy = m * mapSize + m;
-        const frequency = frequencies[ xy ];
-        maxFrequency = Math.max(maxFrequency, frequency);
-    }
-
-    let canvas = document.createElement('canvas');
-    let ctx = canvas.getContext('2d');
-    ctx.canvas.width = ctx.canvas.height = mapSize;
-
-    const { width: w, height: h } = ctx.canvas;
-    ctx.fillStyle = rgb255String( appleCrayonColorRGB255('snow') );
-    ctx.fillRect(0, 0, w, h);
-
-    for (let i = 0; i < w; i++) {
-        for(let j = 0; j < h; j++) {
-
-            const ij = i * w + j;
-
-            let interpolant;
-
-            if (frequencies[ ij ] > maxFrequency) {
-                console.log(`ERROR! At i ${ i } j ${ j } frequencies ${ frequencies[ ij ] } should NOT exceed the max ${ maxFrequency }`);
-                interpolant = maxFrequency / maxFrequency;
-            } else {
-                interpolant = frequencies[ ij ] / maxFrequency;
-            }
-
-            ctx.fillStyle = Globals.colorMapManager.retrieveRGB255String('bintu_et_al', interpolant);
-            ctx.fillRect(i, j, 1, 1);
-        }
-    }
-
-    console.timeEnd(`Create ${ mapSize } x ${ mapSize } contact map`);
-
-    return canvas;
-
-};
-
-export const getEnsembleDistanceMapCanvas = ensemble => {
-
-    const ensembleList = Object.values(ensemble);
-
-    console.time(`getEnsembleDistanceMapCanvas. ${ ensembleList.length } traces.`);
-
-    let mapSize = Globals.ensembleManager.maximumSegmentID;
-
-    let averageDistances = new Array(mapSize * mapSize);
-    for (let d = 0; d < averageDistances.length; d++) averageDistances[ d ] = 0.0;
-
-    for (let trace of ensembleList) {
-
-        const { distanceMapArray, minDistance, maxDistance } = createDistanceArray(trace);
-
-        let { vertices } = trace.geometry;
-        let { length } = vertices;
-        let { segmentList } = trace;
-        for (let i = 0; i < length; i++) {
-
-            const i_segmentIDIndex = segmentList[ i ].segmentID - 1;
-            for (let j = 0; j < length; j++) {
-
-                const j_segmentIDIndex = segmentList[ j ].segmentID - 1;
-
-                const ij =  i_segmentIDIndex * mapSize + j_segmentIDIndex;
-                const ji =  j_segmentIDIndex * mapSize + i_segmentIDIndex;
-
-                // Incrementally build running average
-                averageDistances[ ij ] = averageDistances[ ij ] + (distanceMapArray[ ij ] - averageDistances[ ij ]) / (1 + i);
-
-                averageDistances[ ji ] = averageDistances[ ij ];
-
-            } // for (trace.length)
-
-        } // for (trace.length)
-
-    } // for (trace)
-
-    let minAverageDistance = Number.POSITIVE_INFINITY;
-    let maxAverageDistance = Number.NEGATIVE_INFINITY;
-    for (let averageDistance of averageDistances) {
-        minAverageDistance = Math.min(minAverageDistance, averageDistance);
-        maxAverageDistance = Math.max(maxAverageDistance, averageDistance);
-    }
-
-    console.timeEnd(`getEnsembleDistanceMapCanvas. ${ ensembleList.length } traces.`);
-
-    return createAverageDistanceMap(averageDistances, minAverageDistance, maxAverageDistance);
-
-};
-
-export const getTraceDistanceMapCanvas = trace => {
-
-    const { distanceMapArray, minDistance, maxDistance } = createDistanceArray(trace);
-
-    return createAverageDistanceMap(distanceMapArray, minDistance, maxDistance);
-
-};
-
-const createDistanceArray = trace => {
-
-    let mapSize = Globals.ensembleManager.maximumSegmentID;
-
-    let distanceMapArray = new Array(mapSize * mapSize);
-    for (let d = 0; d < distanceMapArray.length; d++) distanceMapArray[ d ] = -1;
-
-    let maxDistance = Number.NEGATIVE_INFINITY;
-    let minDistance = Number.POSITIVE_INFINITY;
-
-    let { vertices } = trace.geometry;
-    let { segmentList } = trace;
-    let { length } = vertices;
-
-    let exclusionSet = new Set();
-
-    for (let i = 0; i < length; i++) {
-
-        const i_segmentIDIndex = segmentList[ i ].segmentID - 1;
-
-        const xy_diagonal = i_segmentIDIndex * mapSize + i_segmentIDIndex;
-        distanceMapArray[ xy_diagonal ] = 0;
-
-        exclusionSet.add(i);
-
-        for (let j = 0; j < length; j++) {
-
-            if (!exclusionSet.has(j)) {
-
-                const distance = vertices[ i ].distanceTo(vertices[ j ]);
-
-                const j_segmentIDIndex = segmentList[ j ].segmentID - 1;
-
-                const ij =  i_segmentIDIndex * mapSize + j_segmentIDIndex;
-                const ji =  j_segmentIDIndex * mapSize + i_segmentIDIndex;
-
-                distanceMapArray[ ij ] = distanceMapArray[ ji ] = distance;
-
-                maxDistance = Math.max(maxDistance, distance);
-                minDistance = Math.min(minDistance, distance);
-
-            }
-
-        } // for (j)
-
-    }
-
-    return { distanceMapArray, minDistance, maxDistance };
-
-};
-
-const createAverageDistanceMap = (averageDistances, minAverageDistance, maxAverageDistance) => {
-
-    let canvas = document.createElement('canvas');
-    let ctx = canvas.getContext('2d');
-    ctx.canvas.width = ctx.canvas.height = Globals.ensembleManager.maximumSegmentID;
-
-    const { width: w, height: h } = ctx.canvas;
-    ctx.fillStyle = rgb255String( appleCrayonColorRGB255('magnesium') );
-    ctx.fillRect(0, 0, w, h);
-
-    for (let i = 0; i < w; i++) {
-        for(let j = 0; j < h; j++) {
-
-            const ij = i * w + j;
-
-            let interpolant;
-
-            if (i === j && -1 !== averageDistances[ ij ]) {
-                interpolant = 0;
-                ctx.fillStyle = Globals.colorMapManager.retrieveRGB255String('juicebox_default', interpolant);
-                ctx.fillRect(i, j, 1, 1);
-            } else if (-1 === averageDistances[ ij ]) {
-                // do nothing
-                // console.log('ignore');
-            } else {
-                interpolant = (averageDistances[ ij ] - minAverageDistance) / (maxAverageDistance - minAverageDistance);
-                ctx.fillStyle = Globals.colorMapManager.retrieveRGB255String('juicebox_default', interpolant);
-                ctx.fillRect(i, j, 1, 1);
-            }
-
-        }
-
-    }
-
-    return canvas;
-
-};
-
-const kdBushConfiguratorWithTrace = trace => {
-
-    return {
-        idList: trace.geometry.vertices.map((vertex, index) => index),
-        points: trace.geometry.vertices,
-        getX: pt => pt.x,
-        getY: pt => pt.y,
-        getZ: pt => pt.z,
-        nodeSize: 64,
-        ArrayType: Float64Array,
-        axisCount: 3
-    }
 
 };
